chore(components/router): rename router to ingress
