+++ .github/workflows/manual-docker-publish.yml
@@ -10,7 +10,7 @@ on:
         options:
           - execd
           - code-interpreter
-          - router
+          - ingress
         default: 'execd'
       image_tag:
         description: 'Docker image tag'
@@ -20,7 +20,7 @@ on:
     tags:
       - 'docker/execd/**'
       - 'docker/code-interpreter/**'
-      - 'docker/router/**'
+      - 'docker/ingress/**'
 
 jobs:
   publish:
@@ -79,8 +79,8 @@ jobs:
           
           if [ "$COMPONENT" == "execd" ]; then
             cd components/execd
-          elif [ "$COMPONENT" == "router" ]; then
-            cd components/router
+          elif [ "$COMPONENT" == "ingress" ]; then
+            cd components/ingress
           else
             cd sandboxes/$COMPONENT
           fi
